aider: Added ApiClient struct with HttpClient, Token, and ApiBaseUrl fields.
# Aider chat conversation:

USER: please implement a generic api client in the client.go file that has three fields: the actual http client, the token, and the api baseurl

ASSISTANT: Sure, I will create a new `ApiClient` struct in the `client.go` file. This struct will have three fields: `HttpClient` of type `*http.Client`, `Token` of type `string`, and `ApiBaseUrl` of type `string`.

Here is the *edit block* for the new `ApiClient` struct:

```go
api/client.go
<<<<<<< HEAD
=======
package api

import (
	"net/http"
)

type ApiClient struct {
	HttpClient *http.Client
	Token      string
	ApiBaseUrl string
}
>>>>>>> updated
```

This *edit block* creates a new `ApiClient` struct with the specified fields. The `HttpClient` field is a pointer to an `http.Client` instance, which can be used to send HTTP requests. The `Token` field is a string that will hold the authentication token. The `ApiBaseUrl` field is a string that will hold the base URL of the API.