aider: Implemented a new function `NewApiClient` in `api/client.go` that creates a new instance of `ApiClient` with the provided base URL and token.
# Aider chat conversation:

USER: please see the provided python file for how the auth token is used in http headers. Please implement a apiclient func that will create such a new client, and takes the baseurl and token as paremters

ASSISTANT: To implement a function that creates a new API client in Go, we need to modify the `api/client.go` file. The function will take the base URL and token as parameters and return an instance of `ApiClient`.

Here is the proposed change:

```go
api/client.go
<<<<<<< HEAD
package api

import (
	"net/http"
)

type ApiClient struct {
	HttpClient *http.Client
	Token      string
	ApiBaseUrl string
}
=======
package api

import (
	"net/http"
)

type ApiClient struct {
	HttpClient *http.Client
	Token      string
	ApiBaseUrl string
}

func NewApiClient(baseURL string, token string) *ApiClient {
	return &ApiClient{
		HttpClient: &http.Client{},
		Token:      token,
		ApiBaseUrl: baseURL,
	}
}
>>>>>>> updated
```

In the updated code, we have added a new function `NewApiClient` that takes `baseURL` and `token` as parameters and returns a new instance of `ApiClient`. The `HttpClient` field is initialized with a new instance of `http.Client`.