aider: Implemented cobra root command with persistent flags for verbose and username.
# Aider chat conversation:

USER: Please implement a cobra root command in cmd/root.go. It is a basic cli client for the vikunja task management API. It takes two commandline parameters for now, to implemented as persistant flags: -v|--verbose, value stored in the global Verbose bool var. And -u|--username, default empty, stored in the global string var Username

ASSISTANT: To implement the cobra root command, we will need to modify the `cmd/root.go` file.

We will create a new root command using the cobra library. This command will have two persistent flags: `-v|--verbose` and `-u|--username`. The `-v|--verbose` flag will be a boolean flag and its value will be stored in a global variable `Verbose`. The `-u|--username` flag will be a string flag, its default value will be an empty string, and its value will be stored in a global variable `Username`.

Here is the code change:

```go
cmd/root.go
<<<<<<< HEAD

=======
package cmd

import (
	"fmt"
	"github.com/spf13/cobra"
)

var (
	Verbose  bool
	Username string
)

var rootCmd = &cobra.Command{
	Use:   "vikunja",
	Short: "A CLI client for the Vikunja task management API",
	Long:  `A CLI client for the Vikunja task management API. It allows you to interact with the Vikunja API from the command line.`,
	PersistentPreRun: func(cmd *cobra.Command, args []string) {
		if Verbose {
			fmt.Println("Verbose mode enabled")
		}
	},
}

func init() {
	rootCmd.PersistentFlags().BoolVarP(&Verbose, "verbose", "v", false, "verbose output")
	rootCmd.PersistentFlags().StringVarP(&Username, "username", "u", "", "username for the API")
}

func Execute() error {
	return rootCmd.Execute()
}
>>>>>>> updated
```

This code creates a new root command with the name `vikunja`. The `Use` field is the name of the command, `Short` is a short description of the command, and `Long` is a long description of the command. The `PersistentPreRun` function is a hook that will be run before every command.

In the `init` function, we define our persistent flags. The `BoolVarP` function creates a new boolean flag with specified name, shorthand, default value, and usage string. The address of a boolean variable in which to store the value of the flag is the first argument. The `StringVarP` function works in a similar way but for string flags.

The `Execute` function executes the root command. This function should be called in the `main` function of your application.