aider: Added the `CreateTask` method to the `ApiClient` struct in `api/tasks.go`.
# Aider chat conversation:

ASSISTANT: To implement the task creation method, we need to modify the `api/tasks.go` file.

The new method will be named `CreateTask` and will take two parameters: the project ID and the task to be created. It will use the `Put` method from the `ApiClient` struct to send a PUT request to the `/projects/{id}/tasks` endpoint. The task will be serialized to JSON and included in the request body. The response will be deserialized into a `Task` struct and returned.

Here is the *edit block* for the new method:

```go
api/tasks.go
<<<<<<< HEAD
/*
TODO: implement this api method to create a task
*/
=======
func (client *ApiClient) CreateTask(projectID int, task Task) (Task, error) {
	taskJson, err := json.Marshal(task)
	if err != nil {
		return Task{}, err
	}
	response, err := client.Put("/projects/"+strconv.Itoa(projectID)+"/tasks", string(taskJson))
	if err != nil {
		return Task{}, err
	}
	var createdTask Task
	err = json.Unmarshal([]byte(response), &createdTask)
	if err != nil {
		return Task{}, err
	}
	return createdTask, nil
}
>>>>>>> updated
```

This *edit block* adds the `CreateTask` method to the `ApiClient` struct in the `api/tasks.go` file.

USER: InvalidEditBlock: edit failed!

api/tasks.go does not contain the *exact sequence* of HEAD lines you specified.
Try again.
DO NOT skip blank lines, comments, docstrings, etc!
The HEAD block needs to be EXACTLY the same as the lines in api/tasks.go with nothing missing!

api/tasks.go does not contain these 3 exact lines in a row:
```
/*
TODO: implement this api method to create a task
*/
```

ASSISTANT: I apologize for the confusion. It seems I made a mistake in identifying the exact location of the code to be replaced. Let's try again.

We will add the `CreateTask` method at the end of the `api/tasks.go` file. Here is the *edit block* for the new method:

```go
api/tasks.go
<<<<<<< HEAD

=======
func (client *ApiClient) CreateTask(projectID int, task Task) (Task, error) {
	taskJson, err := json.Marshal(task)
	if err != nil {
		return Task{}, err
	}
	response, err := client.Put("/projects/"+strconv.Itoa(projectID)+"/tasks", string(taskJson))
	if err != nil {
		return Task{}, err
	}
	var createdTask Task
	err = json.Unmarshal([]byte(response), &createdTask)
	if err != nil {
		return Task{}, err
	}
	return createdTask, nil
}
>>>>>>> updated
```

This *edit block* adds the `CreateTask` method to the `ApiClient` struct in the `api/tasks.go` file.