aider: Include the `DoneAt` field in the output of the `rootCmd` listing if it is set to an actual datetime.
# Aider chat conversation:

USER: Please see if the tasks struct has a filed showing the date of completion.

ASSISTANT: The `Task` struct in the `api/types.go` file has a field named `DoneAt` of type `time.Time` which likely represents the date and time of task completion. Here is the relevant part of the `Task` struct:

```go
type Task struct {
	...
	Done           bool           `json:"done"`
	DoneAt         time.Time      `json:"done_at"`
	...
}
```

The `DoneAt` field is likely set when the `Done` field is set to `true`. However, without seeing the code that manipulates these fields, I can't confirm this. If you want to check or modify the code that sets these fields, you would need to add the relevant files to the chat.

USER: Please modify the output of the `rootCmd`listing to show this done date if it is set to an actual datetime

ASSISTANT: To include the `DoneAt` field in the output of the `rootCmd` listing, we need to modify the `cmd/root.go` file. Specifically, we need to change the `Run` function of `rootCmd` to check if the `DoneAt` field is set to an actual datetime and, if so, print it.

Here is the *edit block* for the change:

```go
cmd/root.go
<<<<<<< HEAD
		if Verbose {
			formattedTasks, _ := json.MarshalIndent(tasks, "", "  ")
			fmt.Println(string(formattedTasks))
		} else {
			for _, task := range tasks {
				fmt.Printf("%d:  %s (Urgency: %.3f)\n", task.ID, task.Title, task.Urgency)
				if task.Description != "" {
					fmt.Printf("Description: %s\n", task.Description)
				}
				if !task.DueDate.IsZero() {
					fmt.Printf("Due Date: %s\n", task.DueDate.Format("2006-01-02"))
				}
				// fmt.Println()
			}
		}
=======
		if Verbose {
			formattedTasks, _ := json.MarshalIndent(tasks, "", "  ")
			fmt.Println(string(formattedTasks))
		} else {
			for _, task := range tasks {
				fmt.Printf("%d:  %s (Urgency: %.3f)\n", task.ID, task.Title, task.Urgency)
				if task.Description != "" {
					fmt.Printf("Description: %s\n", task.Description)
				}
				if !task.DueDate.IsZero() {
					fmt.Printf("Due Date: %s\n", task.DueDate.Format("2006-01-02"))
				}
				if task.Done {
					fmt.Printf("Done At: %s\n", task.DoneAt.Format("2006-01-02 15:04:05"))
				}
				// fmt.Println()
			}
		}
>>>>>>> updated
```

This change adds a check for `task.Done` and, if `true`, prints the `DoneAt` field in the format "YYYY-MM-DD HH:MM:SS".