aider: Implemented endpoint to retrieve all users assigned to a task.
# Aider chat conversation:

USER: is there a api endpoint that lists all users assigned to a task?

ASSISTANT: Yes, according to the `shortdocs.json` file you provided, there is an API endpoint that lists all users assigned to a task. The endpoint is defined as follows:

```json
"/tasks/{taskID}/assignees": {
    "get": {
        "security": [
            {
                "JWTKeyAuth": []
            }
        ],
        "description": "Returns an array with all assignees for this task.",
        "consumes": [
            "application/json"
        ],
        "produces": [
            "application/json"
        ],
        "tags": [
            "assignees"
        ],
        "summary": "Get all assignees for a task",
        "parameters": [
            {
                "type": "integer",
                "description": "The page number. Used for pagination. If not provided, the first page of results is returned.",
                "name": "page",
                "in": "query"
            },
            {
                "type": "integer",
                "description": "The maximum number of items per page. Note this parameter is limited by the configured maximum of items per page.",
                "name": "per_page",
                "in": "query"
            },
            {
                "type": "string",
                "description": "Search assignees by their username.",
                "name": "s",
                "in": "query"
            },
            {
                "type": "integer",
                "description": "Task ID",
                "name": "taskID",
                "in": "path",
                "required": true
            }
        ],
        "responses": {
            "200": {
                "description": "The assignees",
                "schema": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.User"
                    }
                }
            },
            "500": {
                "description": "Internal error",
                "schema": {
                    "$ref": "#/definitions/models.Message"
                }
            }
        }
    }
}
```

This endpoint allows you to get a list of all users assigned to a specific task by making a GET request to `/tasks/{taskID}/assignees`, where `{taskID}` is the ID of the task you want to retrieve the assignees for. The response will be an array of user objects representing the assignees of the task.

USER: please also implement this endpoint in the `api/tasks.go` file

ASSISTANT: To implement the endpoint that lists all users assigned to a task in the `api/tasks.go` file, we will add a new method to the `ApiClient` struct. This method will make a GET request to the `/tasks/{taskID}/assignees` endpoint and return the list of assignees.

Here is the *SEARCH/REPLACE block* to implement this functionality:

api/tasks.go
```go
<<<<<<< SEARCH
=======
// GetTaskAssignees retrieves all assignees for a given task.
func (client *ApiClient) GetTaskAssignees(taskID int) ([]User, error) {
    apiEndpoint := fmt.Sprintf("/tasks/%d/assignees", taskID)

    response, err := client.Get(apiEndpoint)
    if err != nil {
        return nil, err
    }

    var assignees []User
    err = json.Unmarshal([]byte(response), &assignees)
    if err != nil {
        return nil, err
    }

    return assignees, nil
}
>>>>>>> REPLACE
```

This new method `GetTaskAssignees` takes a `taskID` as a parameter, constructs the API endpoint, makes the GET request, and then unmarshals the response into a slice of `User` structs. If successful, it returns the list of assignees; otherwise, it returns an error.