aider: Added 'assigned' subcommand to list assignees of a task.
# Aider chat conversation:

USER: I'd like to implement to new sub commands to handle assignees to a task. First, let'd use the `assignees` function to list who is assigned to a task. the sub command should be called 'assigned', and take  one param, the task id

ASSISTANT: To implement the new subcommand 'assigned' to list who is assigned to a task, we need to modify `cmd/task.go`. We will add a new `cobra.Command` for the 'assigned' subcommand and implement its `Run` function to use `ApiClient.GetTaskAssignees` to retrieve and display the assignees for the given task ID.

Here is the *SEARCH/REPLACE block* for adding the 'assigned' subcommand:

cmd/task.go
```go
<<<<<<< SEARCH
func init() {
	newCmd.Flags().StringP("due", "d", "", "Due date for the task")
	rootCmd.AddCommand(newCmd)

	rootCmd.AddCommand(doneCmd)
	rootCmd.AddCommand(editCmd)
}
=======
var assignedCmd = &cobra.Command{
	Use:   "assigned",
	Short: "List assignees of a task",
	Long:  `List all the assignees assigned to a task using the provided task ID.`,
	Args:  cobra.ExactArgs(1),
	Run: func(cmd *cobra.Command, args []string) {
		taskID, err := strconv.Atoi(args[0])
		if err != nil {
			fmt.Println("Error converting task ID:", err)
			return
		}
		assignees, err := ApiClient.GetTaskAssignees(taskID)
		if err != nil {
			fmt.Println("Error getting assignees for task:", err)
			return
		}
		for _, assignee := range assignees {
			fmt.Printf("ID: %d, Username: %s, Name: %s\n", assignee.ID, assignee.Username, assignee.Name)
		}
	},
}

func init() {
	newCmd.Flags().StringP("due", "d", "", "Due date for the task")
	rootCmd.AddCommand(newCmd)

	rootCmd.AddCommand(doneCmd)
	rootCmd.AddCommand(editCmd)
	rootCmd.AddCommand(assignedCmd)
}
>>>>>>> REPLACE
```

This block of code adds a new `cobra.Command` named `assignedCmd` and registers it in the `init` function. The `Run` function of `assignedCmd` converts the first argument to an integer to use as the task ID, calls `ApiClient.GetTaskAssignees` to get the list of assignees, and then prints out the details of each assignee.